refactor: move common methods to diff. pkg
